Refactored to separate file
