+++ lib/src/platform.dart
@@ -140,138 +140,6 @@ PlatformTarget platform(BuildContext context) {
   }
 }
 
-Future<T?> showPlatformDialog<T>({
-  required BuildContext context,
-  required WidgetBuilder builder,
-  bool? barrierDismissible,
-  RouteSettings? routeSettings,
-  bool useRootNavigator = true,
-  bool materialUseSafeArea = true,
-  Color? materialBarrierColor = Colors.black54,
-  String? barrierLabel,
-}) {
-  if (isMaterial(context)) {
-    return showDialog<T>(
-      context: context,
-      builder: builder,
-      barrierDismissible: barrierDismissible ?? true,
-      routeSettings: routeSettings,
-      useRootNavigator: useRootNavigator,
-      useSafeArea: materialUseSafeArea,
-      //child: , deprecated
-      barrierColor: materialBarrierColor,
-      barrierLabel: barrierLabel,
-    );
-  } else {
-    return showCupertinoDialog<T>(
-      context: context,
-      builder: builder,
-      routeSettings: routeSettings,
-      useRootNavigator: useRootNavigator,
-      barrierDismissible: barrierDismissible ?? false,
-      barrierLabel: barrierLabel,
-    );
-  }
-}
-
-abstract class _BaseData {
-  _BaseData({
-    this.anchorPoint,
-  });
-
-  final Offset? anchorPoint;
-}
-
-class MaterialModalSheetData extends _BaseData {
-  final Color? backgroundColor;
-  final double? elevation;
-  final ShapeBorder? shape;
-  final bool isScrollControlled;
-  final bool useRootNavigator;
-  final Clip? clipBehavior;
-  final Color? barrierColor;
-  final bool enableDrag;
-  final bool isDismissible;
-  final RouteSettings? routeSettings;
-  final AnimationController? transitionAnimationController;
-  final BoxConstraints? constraints;
-
-  MaterialModalSheetData({
-    super.anchorPoint,
-    this.backgroundColor,
-    this.elevation,
-    this.shape,
-    this.isScrollControlled = false,
-    this.useRootNavigator = false,
-    this.clipBehavior,
-    this.barrierColor,
-    this.enableDrag = false,
-    this.isDismissible = false,
-    this.routeSettings,
-    this.transitionAnimationController,
-    this.constraints,
-  });
-}
-
-class CupertinoModalSheetData extends _BaseData {
-  final ImageFilter? imageFilter;
-  final bool? semanticsDismissible;
-  final bool useRootNavigator;
-  final Color barrierColor;
-  final bool barrierDismissible;
-  final RouteSettings? routeSettings;
-
-  CupertinoModalSheetData({
-    super.anchorPoint,
-    this.imageFilter,
-    this.semanticsDismissible,
-    this.useRootNavigator = true,
-    this.barrierColor = _kModalBarrierColor,
-    this.barrierDismissible = true,
-    this.routeSettings,
-  });
-}
-
-/// Displays either the showModalBottomSheet for material
-/// or showCupertinoModalPopup for cupertino
-Future<T?> showPlatformModalSheet<T>({
-  required BuildContext context,
-  required WidgetBuilder builder,
-  MaterialModalSheetData? material,
-  CupertinoModalSheetData? cupertino,
-}) {
-  if (isMaterial(context)) {
-    return showModalBottomSheet<T>(
-      context: context,
-      builder: builder,
-      backgroundColor: material?.backgroundColor,
-      elevation: material?.elevation,
-      shape: material?.shape,
-      isScrollControlled: material?.isScrollControlled ?? false,
-      useRootNavigator: material?.useRootNavigator ?? false,
-      clipBehavior: material?.clipBehavior,
-      barrierColor: material?.barrierColor,
-      enableDrag: material?.enableDrag ?? true,
-      isDismissible: material?.isDismissible ?? true,
-      routeSettings: material?.routeSettings,
-      transitionAnimationController: material?.transitionAnimationController,
-      constraints: material?.constraints,
-      anchorPoint: material?.anchorPoint,
-    );
-  } else {
-    return showCupertinoModalPopup<T>(
-        context: context,
-        builder: builder,
-        filter: cupertino?.imageFilter,
-        semanticsDismissible: cupertino?.semanticsDismissible,
-        useRootNavigator: cupertino?.useRootNavigator ?? true,
-        barrierColor: cupertino?.barrierColor ?? _kModalBarrierColor,
-        barrierDismissible: cupertino?.barrierDismissible ?? true,
-        routeSettings: cupertino?.routeSettings,
-        anchorPoint: cupertino?.anchorPoint);
-  }
-}
-
 abstract class CustomBuilder<T> {}
 
 typedef PlatformTargetBuilder<T> = Widget Function(
